Allowing for nested brackets
Including a quick formalization of the currently allowed language for
padatious in EBNF.
By this patch it is allowed to nest brackets and the results will be as
expected. This removes parsing 1 ( 2 3 ) 4 to 1(23)4 but instead results
in 1234 as one might expect. Still, only (..|) or (|..) results in empty
subsenteces being allowed.